refactor(console): apply new layout for org settings
