+++ packages/console/src/pages/Organizations/TemplateTable/index.module.scss
@@ -4,10 +4,6 @@
   margin-top: _.unit(3);
 }
 
-.table {
-  margin-top: _.unit(3);
-}
-
 .spinner {
   margin: _.unit(2) 0;
 }
