+++ packages/console/src/pages/Organizations/index.module.scss
@@ -1,5 +1,5 @@
 @use '@/scss/underscore' as _;
 
 .tabs {
-  margin: _.unit(4) 0;
+  margin-top: _.unit(4);
 }
