+++ packages/phrases/src/locales/ko/translation/admin-console/organizations.ts
@@ -17,14 +17,12 @@ const organizations = {
   /** UNTRANSLATED */
   organization_description_placeholder: 'A brief description of the organization.',
   /** UNTRANSLATED */
-  access_control: 'Access control',
-  /** UNTRANSLATED */
-  access_control_description:
-    'Authorization in a multi-tenancy applications is often designed to make sure that tenant isolation is maintained throughout an application and that tenants can access only their own resources.',
-  /** UNTRANSLATED */
   organization_permission: 'Organization permission',
   /** UNTRANSLATED */
   organization_permission_other: 'Organization permissions',
+  /** UNTRANSLATED */
+  organization_permission_description:
+    'Organization permission refers to the authorization to access a resource in the context of organization. An organization permission should be represented as a meaningful string, also serving as the name and unique identifier.',
   /** UNTRANSLATED */
   organization_permission_delete_confirm:
     'If this permission is deleted, all organization roles including this permission will lose this permission, and users who had this permission will lose the access granted by it.',
@@ -38,6 +36,9 @@ const organizations = {
   organization_role: 'Organization role',
   /** UNTRANSLATED */
   organization_role_other: 'Organization roles',
+  /** UNTRANSLATED */
+  organization_role_description:
+    'Organization role is a grouping of permissions that can be assigned to users. The permissions must come from the predefined organization permissions.',
   /** UNTRANSLATED */
   organization_role_delete_confirm:
     'Doing so will remove the permissions associated with this role from the affected users and delete the relations among organization roles, members in the organization, and organization permissions.',
